installing flutter map
