+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: report_it
+name: reportit
 description: A new Flutter project.
 
 # The following defines the version and build number for your application.
@@ -15,6 +15,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_map: ^0.1.4
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
